Pull item cardview
Added the cardview xml file. It will be used to hold the information from the repo regarding its pulls. Added a new Linear Lahoutto the repo pull list to display numbers of pull opened and closed.  Changed the height property in the gitrepo item card to wrap_content, so the view will use more space when needed.