Item data model serialization
Added serialization annotations for the ItemDataModel in order to use it to parse json serializaded objects from retrofit.  Added a Constraint Layout to content main, placed a textview and a progress bar to inform user for connection uses or refresh. Refactored the GitRepoDataModel, to ItemDataModel to better suit the JSON structure from the api.